+++ mod-source-record-manager-server/src/main/java/org/folio/services/mappers/processor/parameters/MappingParameters.java
@@ -3,10 +3,11 @@ package org.folio.services.mappers.processor.parameters;
 import org.apache.commons.collections4.list.UnmodifiableList;
 import org.folio.rest.jaxrs.model.ClassificationType;
 import org.folio.rest.jaxrs.model.ContributorNameType;
+import org.folio.rest.jaxrs.model.ElectronicAccessRelationship;
 import org.folio.rest.jaxrs.model.ContributorType;
 import org.folio.rest.jaxrs.model.IdentifierType;
-import org.folio.rest.jaxrs.model.InstanceType;
 import org.folio.rest.jaxrs.model.InstanceFormat;
+import org.folio.rest.jaxrs.model.InstanceType;
 
 import java.util.List;
 
@@ -18,6 +19,7 @@ public class MappingParameters {
   private UnmodifiableList<IdentifierType> identifierTypes;
   private UnmodifiableList<ClassificationType> classificationTypes;
   private UnmodifiableList<InstanceType> instanceTypes;
+  private UnmodifiableList<ElectronicAccessRelationship> electronicAccessRelationship;
   private UnmodifiableList<InstanceFormat> instanceFormats;
   private UnmodifiableList<ContributorType> contributorTypes;
   private UnmodifiableList<ContributorNameType> contributorNameTypes;
@@ -46,6 +48,15 @@ public class MappingParameters {
 
   public MappingParameters withInstanceTypes(List<InstanceType> instanceTypes) {
     this.instanceTypes = new UnmodifiableList<>(instanceTypes);
+    return this;
+  }
+
+  public List<ElectronicAccessRelationship> getElectronicAccessRelationships() {
+    return electronicAccessRelationship;
+  }
+
+  public MappingParameters withElectronicAccessRelationships(List<ElectronicAccessRelationship> electronicAccessRelationship) {
+    this.electronicAccessRelationship = new UnmodifiableList<>(electronicAccessRelationship);
     return this;
   }