+++ mod-source-record-manager-server/src/test/java/org/folio/rest/impl/AbstractRestTest.java
@@ -75,6 +75,7 @@ public abstract class AbstractRestTest {
   private static final String INSTANCE_FORMATS_URL = "/instance-formats?limit=500";
   private static final String CONTRIBUTOR_TYPES_URL = "/contributor-types?limit=500";
   private static final String CONTRIBUTOR_NAME_TYPES_URL = "/contributor-name-types?limit=500";
+  private static final String ELECTRONIC_ACCESS_URL = "/electronic-access-relationships?limit=500";
   protected static final String FILES_PATH = "src/test/resources/org/folio/rest/files.sample";
   protected static final String RECORD_PATH = "src/test/resources/org/folio/rest/record.json";
   protected static final String SNAPSHOT_SERVICE_URL = "/source-storage/snapshots";
@@ -199,6 +200,8 @@ public abstract class AbstractRestTest {
     WireMock.stubFor(WireMock.get(INSTANCE_TYPES_URL)
       .willReturn(WireMock.okJson(new JsonObject().toString())));
     WireMock.stubFor(WireMock.get(CLASSIFICATION_TYPES_URL)
+      .willReturn(WireMock.okJson(new JsonObject().toString())));
+    WireMock.stubFor(WireMock.get(ELECTRONIC_ACCESS_URL)
       .willReturn(WireMock.okJson(new JsonObject().toString())));
     WireMock.stubFor(WireMock.get(INSTANCE_FORMATS_URL)
       .willReturn(WireMock.okJson(new JsonObject().toString())));