+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 2019-09-09 v1.7.1
  * Added instance-type settings loading for mapping
+ * Added electronic access relationships loading for mapping
  * Added classification settings loading for mapping
  * Added instance-format settings loading for mapping
  * Added contributor name types setting loading for mapping